Edits at home
There are now two directories in the annotated section, one for Linux
and one for MacOSX assembly. The files in these directories are also
heavily annotated, explaining what I did wrong previously and what each
step is for.

The Makefile has been edited to make using a different directory of
sources depending on the OS.