ft_atoi working for Mac
On the way to getting ft_atoi_base.s, I've recreated ft_atoi with
assembly. It's passing tests at the moment, although it could do with
testing with a couple more edge cases in future.

Next step is to recreate it for linux (shouldn't be too difficult), and
then to move on to completing ft_atoi_base.s (which probably will be
very difficult).