Annotations removed on submission files
Files have been taken out of the annotation folder (only MacOSX) in
preparation to submit. Makefile edited to incorporate criterion test
files which still need to be fully completed. Makefile now also compiles
into a static library "libasm.a", and requires the "test" or "bonustest"
instruction to create a test executable.

To-do
	- More criterion testing, for every function created.
	- Separate header file into two, one for bonus and one for
	mandatory sections.