Created more tests, fixed strcpy
ft_strcpy had a pretty awful error: a null-terminator was not being
added. Thankfully it was found in the course of testing and has now been
fixed.

A few more tests have been added and I've begun on the bonus tests now.
Test sources should be placed in a separate directory and I'll get
around to doing that soon.